+++ src/decorators.ts
@@ -1,4 +1,5 @@
 import type {ComponentOptions, ComponentType} from './component';
+import type {SystemGroup, SystemType} from './system';
 import type {Type} from './type';
 
 interface PropOptions<JSType> {
@@ -21,16 +22,56 @@ export function field<JSType>(
 
 export const componentTypes: ComponentType<any>[] = [];
 
-export function component(constructor: ComponentType<any>): void;
+/**
+ * Declare this class as a component type that will be automatically added to any new world.
+ * @param componentClass The component class.
+ */
+export function component(componentClass: ComponentType<any>): void;
+
+/**
+ * Declare this class as a component type that will be automatically added to any new world.
+ * @param options The options to apply to the component type.
+ */
 export function component(options: ComponentOptions): (constructor: ComponentType<any>) => void;
+
 export function component(arg: ComponentType<any> | ComponentOptions):
-    ((constructor: ComponentType<any>) => void) | void {
+    ((componentClass: ComponentType<any>) => void) | void {
   if (typeof arg === 'function') {
     componentTypes.push(arg);
   } else {
-    return (constructor: ComponentType<any>) => {
-      constructor.options = arg;
-      componentTypes.push(constructor);
+    return (componentClass: ComponentType<any>) => {
+      componentClass.options = arg;
+      componentTypes.push(componentClass);
+    };
+  }
+}
+
+
+export const systemTypes: (SystemType<any> | SystemGroup)[] = [];
+
+/**
+ * Declare this class as a system type that will be automatically added to any new world.  The class
+ * must inherit from System.
+ * @param systemClass The system class.
+ */
+export function system(systemClass: SystemType<any>): void;
+
+/**
+ * Declare this class as a system type that will be automatically added to any new world.  The class
+ * must inherit from System.
+ * @param systemGroup A system group to add the system type to. This system group will also be
+ * automatically added to any new world.
+ */
+export function system(systemGroup: SystemGroup): (constructor: SystemType<any>) => void;
+
+export function system(arg: SystemType<any> | SystemGroup):
+    ((systemClass: SystemType<any>) => void) | void {
+  if (typeof arg === 'function') {
+    systemTypes.push(arg);
+  } else {
+    if (!systemTypes.includes(arg)) systemTypes.push(arg);
+    return (systemClass: SystemType<any>) => {
+      arg.__contents.push(systemClass);
     };
   }
 }