+++ tests/reactivequery.test.ts
@@ -1,6 +1,4 @@
-import {
-  component, ComponentType, componentTypes, field, Query, System, SystemType, Type, World,
-} from '../src';
+import {component, ComponentType, field, Query, System, SystemType, Type, World} from '../src';
 
 
 @component class A {
@@ -68,7 +66,7 @@ class Count extends System {
 }
 
 async function createWorld(...systems: SystemType<System>[]): Promise<World> {
-  return World.create({maxEntities: 100, defs: [componentTypes, systems, Count]});
+  return World.create({maxEntities: 100, defs: [systems, Count]});
 }